+++ tests/conftest.py
@@ -5,6 +5,7 @@ import sofar as sf
 import pyfar as pf
 
 from pyfar import Orientations
+from pyfar import Rotation
 from pyfar import Coordinates
 from pyfar import FrequencyData, TimeData
 import pyfar.classes.filter as fo
@@ -737,7 +738,7 @@ def ups():
     """Used for the creation of Orientation objects with
     `Orientations.from_view_up`.
     """
-    return [[0, 1, 0], [0, -2, 0], [0, 1, 0]]
+    return [[0, 0, 1], [0, 0, 2], [0, 0, 1]]
 
 
 @pytest.fixture()
@@ -753,6 +754,13 @@ def orientations(views, ups):
     """Orientations object uses fixtures `views` and `ups`.
     """
     return Orientations.from_view_up(views, ups)
+
+
+@pytest.fixture()
+def rotation(views, ups):
+    """Rotation object uses fixtures `views` and `ups`.
+    """
+    return Rotation.from_view_up(views, ups)
 
 
 @pytest.fixture()