+++ HISTORY.rst
@@ -21,6 +21,7 @@ Added
 - All plots showing the magnitude response now have the `mode` parameter to specify if the absolute (default), real, or imaginary value of the spectrum is shown. This can also be toggled using the shortcut 'shift+m' in interactive plots (PR #817)
 - `pyfar.signals.files.room_impulse_response` can now return the room impulse response with noise tail (PR #906)
 - Abstract `_LTISystem` base class to better structure inheritance of the pyfar Filter classes (PR #798)
+- Added new `Rotation` class as a long term replacement for `Orientations` (PR #904)
 
 Changed
 ^^^^^^^